Created ObjC classes wrapping textures and framebuffers (beginning work on offscreen rendering of layers). Delegated more responsibility to backing store. Some style and bug fixes.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/quartzcore/trunk@35312 72102866-910b-0410-8b05-ffd578937521

+++ Headers/QuartzCore/CAMediaTimingFunction.h
@@ -40,8 +40,8 @@ extern NSString *const kCAMediaTimingFunctionLinear;
   float _c1x, _c1y;
   float _c2x, _c2y;
   
-  float _coefficientsX[4];
-  float _coefficientsY[4];
+  CGFloat _coefficientsX[4];
+  CGFloat _coefficientsY[4];
 }
 
 + (id) functionWithName:(NSString *)name;
